requirements.txt
fastapi==0.116.1
pydantic==2.11.7
starlette==0.47.3
uvicorn==0.35.0
pydantic_settings==2.10.1
python-dotenv~=1.1
python-multipart==0.0.20
pandas==2.3.2
openai==1.102.0
llmwhisperer-client==2.4.2
